+++ src/main/java/Entities/Target.java
@@ -1,4 +1,4 @@
-package entities;
+package Entities;
 
 import java.util.Date;
