+++ src/main/java/Entities/TargetORM.java
@@ -1,4 +1,4 @@
-package entities;
+package Entities;
 
 import java.util.ArrayList;
