+++ src/main/java/Entities/TargetWeight.java
@@ -1,4 +1,4 @@
-package entities;
+package Entities;
 
 import java.util.ArrayList;
