+++ src/main/java/Entities/Targets.java
@@ -1,4 +1,4 @@
-package entities;
+package Entities;
 
 import java.util.ArrayList;
