+++ src/main/java/Usecases/EditTarget.java
@@ -1,8 +1,7 @@
-package usecases;
+package Usecases;
 
-import entities.Target;
-import entities.TargetORM;
-import entities.Targets;
+import Entities.Target;
+import Entities.Targets;
 
 import java.util.ArrayList;
 import java.util.Date;
