+++ src/main/java/Views/DeleteTargetOptions.java
@@ -1,9 +1,9 @@
-package UI;
+package Views;
 
-import controllers.DeleteController;
-import entities.TargetORM;
-import entities.TargetVolume;
-import entities.TargetWeight;
+import Controllers.DeleteController;
+import Entities.TargetORM;
+import Entities.TargetVolume;
+import Entities.TargetWeight;
 
 import javax.swing.*;
 /**
